backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.18
python-multipart==0.0.6
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
python-dateutil==2.8.2
bcrypt==4.0.1
pyjwt==2.8.0
python-dotenv==1.0.0
torch==2.0.1
textblob==0.17.1
scipy==1.11.3